requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2024.7.4
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distlib==0.3.8
dnspython==2.7.0
fastapi==0.115.12
filelock==3.15.4
h11==0.16.0
idna==3.10
pipenv==2024.0.1
platformdirs==4.2.2
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
scapy==2.6.1
schedule==1.2.2
setuptools==70.3.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.26.3
